server/requirements.txt
uvicorn==0.19.0
fastapi==0.87.0
pydantic~=1.10.2

PyJWT==2.6.0
python-decouple==3.6
passlib~=1.7.4